requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.24.0
aio-pika==9.5.8
pydantic==2.10.0
python-dotenv==1.0.0
scikit-learn==1.8.0
joblib==1.3.2
pytest==7.4.3
httpx==0.27.0
redis==7.4.0
